pyproject.toml
[project]
name = "ai-school-agentic-app"
version = "0.1.0"
description = "Weather + News agent with hand-rolled MCP-backed agent loop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "mcp>=1.27.0",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.24.1",
]

[project.scripts]
ai-school-agent = "ai_school_agentic_app.cli:app"
ai-school-eval = "eval.run_eval:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
